Refactored _padeXX() functs to use sparse or dense matrices.
